feat: update configé

+++ bin/dark_theme.sh
@@ -4,5 +4,7 @@
 starship config palette rose-pine
 
 # kitty
-echo "include dark-theme.auto.conf" > $HOME/.config/kitty/no-preference-theme.auto.conf
-kill -s USR1 $(pidof kitty)
+if [[ "$OSTYPE" != "darwin"* ]]; then
+  echo "include dark-theme.auto.conf" > $HOME/.config/kitty/no-preference-theme.auto.conf
+  kill -s USR1 $(pidof kitty)
+fi
