+++ bin/light_theme.sh
@@ -4,5 +4,7 @@
 starship config palette rose-pine-dawn
 
 # kitty
-echo "include light-theme.auto.conf" > $HOME/.config/kitty/no-preference-theme.auto.conf
-kill -s USR1 $(pidof kitty)
+if [[ "$OSTYPE" != "darwin"* ]]; then
+  echo "include light-theme.auto.conf" > $HOME/.config/kitty/no-preference-theme.auto.conf
+  kill -s USR1 $(pidof kitty)
+fi
